+++ web/static/react/assets/ColdStartInterviewPage-D50I_vJ4.js
@@ -1,1 +1,0 @@
-import{j as r}from"./vendor-react-DCA3L5Xq.js";import{L as t}from"./LegalSkillsProfilePage-uxrD_lex.js";import"./button-COuRSX1r.js";import"./index-Ck1mmUAz.js";import"./vendor-icons-BX2-bB2Y.js";import"./input-Bo3AU9Fz.js";import"./IusPageShell-Bhutex0z.js";import"./api-ByCF5bE6.js";function n(){return r.jsx(t,{})}export{n as ColdStartInterviewPage};