+++ web/static/react/assets/ColdStartInterviewPage-eCKasqCW.js
@@ -1,0 +1,1 @@
+import{j as r}from"./vendor-react-DCA3L5Xq.js";import{L as t}from"./LegalSkillsProfilePage-Cfuyr1r4.js";import"./button-D5QFqFqd.js";import"./index-H0-EOAh9.js";import"./vendor-icons-BX2-bB2Y.js";import"./input-DeEwMjyU.js";import"./IusPageShell-_KnMXfS8.js";import"./api-CxwxUWSo.js";function n(){return r.jsx(t,{})}export{n as ColdStartInterviewPage};