+++ web/static/react/assets/PracticeProfilePage-C9iCNHLJ.js
@@ -1,1 +1,0 @@
-import{j as r}from"./vendor-react-DCA3L5Xq.js";import{L as o}from"./LegalSkillsProfilePage-uxrD_lex.js";import"./button-COuRSX1r.js";import"./index-Ck1mmUAz.js";import"./vendor-icons-BX2-bB2Y.js";import"./input-Bo3AU9Fz.js";import"./IusPageShell-Bhutex0z.js";import"./api-ByCF5bE6.js";function l(){return r.jsx(o,{})}export{l as PracticeProfilePage};